% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_drive.R
\name{box_drive}
\alias{box_drive}
\title{Get path on Box Drive}
\usage{
box_drive(...)
}
\arguments{
\item{...}{\code{character} elements of path, with respect to root of Box Drive.}
}
\value{
\code{character} absolute path on Box Drive.
}
\description{
Get path on Box Drive
}
