% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose.R
\name{dr_box_drive}
\alias{dr_box_drive}
\title{Diagnose Box Drive}
\usage{
dr_box_drive()
}
\value{
Invisible \code{NULL}, called for side effects.
}
\description{
Provides diagnostics on access to local computer's installation of Box Drive.
To reload Box Drive's settings from your computer, use \code{reload = TRUE}.
}
\examples{
  dr_box_drive()
}
