% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose.R
\name{dr_box_drive}
\alias{dr_box_drive}
\title{Diagnose Box Drive}
\usage{
dr_box_drive(reload = TRUE)
}
\arguments{
\item{reload}{\code{logical} indicates to reload Box Drive settings.}
}
\value{
Invisible \code{NULL}, called for side effects.
}
\description{
Diagnose Box Drive
}
\examples{
  dr_box_drive()
}
